// Frontend/src/enums/pcm006.ts
export enum EPcm006Status {
  All = "All",
  Draft = "Draft",
  Edit = "Edit",
  WaitingApproval = "WaitingApproval",
  // Approved = "Approved",
  Rejected = "Rejected",
  Cancelled = "Cancelled",
  WaitingAccountingApproval = "WaitingAccountingApproval",
  WaitingDisbursementDate = "WaitingDisbursementDate",
  Paid = "Paid"
}